Add files via upload
This project was inspired by the amazing presentations and demos showcased at Google I/O. While it may not be perfect, it represents my journey in learning HTML and web development. I've poured my passion and dedication into creating this project, and I'm excited to share it with you all.